+++ test/library_card_file_menu_test.dart
@@ -264,6 +264,7 @@ void main() {
           playerServiceFactory: ({
             required String hwdec,
             required bool enableHardwareAcceleration,
+            required PlayerRendererPreference rendererPreference,
           }) =>
               PlayerService(backend: _CardFileMenuPlayerBackend()),
           mediaProbeBackendFactory: _CardFileMenuProbeBackend.new,